--- src/main/java/ru/mipt/bit/platformer/actors/commands/ActionCommand.java
@@ -1,5 +1,0 @@
-package ru.mipt.bit.platformer.actors.commands;
-
-public interface ActionCommand {
-    void act();
-}
