+++ src/main/java/ru/mipt/bit/platformer/controllers/Controller.java
@@ -1,0 +1,9 @@
+package ru.mipt.bit.platformer.controllers;
+
+import ru.mipt.bit.platformer.controllers.commands.ActionCommand;
+
+import java.util.List;
+
+public interface Controller {
+    List<ActionCommand> getCommands(float deltaTime);
+}
