+++ src/main/java/ru/mipt/bit/platformer/controllers/commands/ActionCommand.java
@@ -1,0 +1,5 @@
+package ru.mipt.bit.platformer.controllers.commands;
+
+public interface ActionCommand {
+    void act();
+}
