+++ src/main/java/ru/mipt/bit/platformer/controllers/commands/ShootCommand.java
@@ -1,7 +1,6 @@
-package ru.mipt.bit.platformer.actors.commands;
+package ru.mipt.bit.platformer.controllers.commands;
 
 import ru.mipt.bit.platformer.objects.tank.Tank;
-import ru.mipt.bit.platformer.util.Direction;
 
 public class ShootCommand implements  ActionCommand {
     private final Tank targetTank;
