+++ src/test/java/ru/mipt/bit/platformer/controllers/PlayerTest.java
@@ -1,10 +1,10 @@
-package ru.mipt.bit.platformer.actors;
+package ru.mipt.bit.platformer.controllers;
 
 import com.badlogic.gdx.math.GridPoint2;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import ru.mipt.bit.platformer.ColliderManager;
-import ru.mipt.bit.platformer.actors.commands.ActionCommand;
+import ru.mipt.bit.platformer.controllers.commands.ActionCommand;
 import ru.mipt.bit.platformer.input.directions.Directions;
 import ru.mipt.bit.platformer.objects.tank.LightDamagedTank;
 import ru.mipt.bit.platformer.objects.tank.Tank;
